requirements/requirements.txt
streamlit
pytesseract
pdfplumber
Pillow
pandas
sentence-transformers
chromadb
opencv-python
pdf2image
